+++ src/projects/project-operations.spec.ts
@@ -116,6 +116,7 @@ describe('Project Operations', () => {
 				primaryLanguage: 'en-GB',
 				supportedLanguages: ['fr-FR', 'de-DE'],
 				deliverySysExclusions: [],
+				color: 'blue',
 			};
 			let project = await client.projects.create(newProject);