+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.1] - 2025-09-19
+### Added
+- export *Component* type with appended metadata
+- export *FormContentType* and forms-specific field types
+- missing types in *EntrySys*: *properties*, *review*, *schedules* and *versionStatus*
+
+### Changed
+- return *Component* type with appended metadata in component operations
+- restrict *EntrySys.metadata* properties to string type only as complex types are not supported by the API and silently break
+- make *EntrySys.uri* optional - present on *entries.get()*, absent on *search()*/*list()*, set in site view via discrete *nodes* operations.
+
 ## [2.4.0] - 2025-06-02
 
 ### Added